feat: various improvements for szuru
