+++ lib/boorus/szurubooru/create_szurubooru_config_page.dart
@@ -29,6 +29,7 @@ class CreateSzurubooruConfigPage extends ConsumerWidget {
         initialBooruConfigProvider.overrideWithValue(config),
       ],
       child: CreateBooruConfigScaffold(
+        hasRatingFilter: true,
         isNewConfig: isNewConfig,
         initialTab: initialTab,
         backgroundColor: backgroundColor,
