+++ lib/boorus/szurubooru/szurubooru_post_details_page.dart
@@ -1,0 +1,145 @@
+// Flutter imports:
+import 'package:flutter/material.dart';
+
+// Package imports:
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+// Project imports:
+import 'package:boorusama/boorus/booru_builder.dart';
+import 'package:boorusama/core/posts/posts.dart';
+import 'package:boorusama/core/tags/tags.dart';
+import 'package:boorusama/core/widgets/widgets.dart';
+import 'package:boorusama/dart.dart';
+import 'package:boorusama/functional.dart';
+import 'post_votes/post_votes.dart';
+import 'szurubooru_post.dart';
+
+class SzurubooruPostDetailsPage extends ConsumerWidget {
+  const SzurubooruPostDetailsPage({
+    super.key,
+    required this.controller,
+    required this.onExit,
+    required this.onPageChanged,
+    required this.posts,
+    required this.initialPage,
+  });
+
+  final List<Post> posts;
+  final PostDetailsController<Post> controller;
+  final void Function(int page) onExit;
+  final void Function(int page) onPageChanged;
+  final int initialPage;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return PostDetailsPageScaffold(
+      posts: posts,
+      initialIndex: initialPage,
+      swipeImageUrlBuilder: defaultPostImageUrlBuilder(ref),
+      onExit: onExit,
+      onPageChangeIndexed: onPageChanged,
+      statsTileBuilder: (context, rawPost) =>
+          castOrNull<SzurubooruPost>(rawPost).toOption().fold(
+                () => const SizedBox.shrink(),
+                (post) => Column(
+                  children: [
+                    const Divider(height: 8, thickness: 0.5),
+                    SimplePostStatsTile(
+                      totalComments: post.commentCount,
+                      favCount: post.favoriteCount,
+                      score: post.score,
+                    ),
+                  ],
+                ),
+              ),
+      tagListBuilder: (context, post) =>
+          castOrNull<SzurubooruPost>(post).toOption().fold(
+                () => const SizedBox.shrink(),
+                (post) => TagsTile(
+                  post: post,
+                  tags: createTagGroupItems(post.tagDetails),
+                  initialExpanded: true,
+                  tagColorBuilder: (tag) => tag.category.darkColor,
+                ),
+              ),
+      toolbar: ValueListenableBuilder(
+        valueListenable: controller.currentPost,
+        builder: (_, rawPost, __) =>
+            castOrNull<SzurubooruPost>(rawPost).toOption().fold(
+                  () => SimplePostActionToolbar(post: rawPost),
+                  (post) => SzurubooruPostActionToolbar(post: post),
+                ),
+      ),
+      fileDetailsBuilder: (context, rawPost) => DefaultFileDetailsSection(
+        post: rawPost,
+        uploaderName: castOrNull<SzurubooruPost>(rawPost)?.uploaderName,
+      ),
+    );
+  }
+}
+
+class SzurubooruPostDetailsDesktopPage extends ConsumerWidget {
+  const SzurubooruPostDetailsDesktopPage({
+    super.key,
+    required this.initialIndex,
+    required this.posts,
+    required this.onExit,
+    required this.onPageChanged,
+    required this.controller,
+  });
+
+  final int initialIndex;
+  final List<Post> posts;
+  final void Function(int index) onExit;
+  final void Function(int page) onPageChanged;
+  final PostDetailsController<Post> controller;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return PostDetailsPageDesktopScaffold(
+      initialIndex: initialIndex,
+      posts: posts,
+      onExit: onExit,
+      onPageChanged: onPageChanged,
+      imageUrlBuilder: defaultPostImageUrlBuilder(ref),
+      statsTileBuilder: (context, rawPost) =>
+          castOrNull<SzurubooruPost>(rawPost).toOption().fold(
+                () => const SizedBox.shrink(),
+                (post) => Column(
+                  children: [
+                    const Divider(height: 8, thickness: 0.5),
+                    SimplePostStatsTile(
+                      totalComments: post.commentCount,
+                      favCount: post.favoriteCount,
+                      score: post.score,
+                    ),
+                  ],
+                ),
+              ),
+      tagListBuilder: (context, post) =>
+          castOrNull<SzurubooruPost>(post).toOption().fold(
+                () => const SizedBox.shrink(),
+                (post) => TagsTile(
+                  post: post,
+                  tags: createTagGroupItems(post.tagDetails),
+                  initialExpanded: true,
+                  tagColorBuilder: (tag) => tag.category.darkColor,
+                ),
+              ),
+      toolbarBuilder: (context, post) => ValueListenableBuilder(
+        valueListenable: controller.currentPost,
+        builder: (_, rawPost, __) =>
+            castOrNull<SzurubooruPost>(rawPost).toOption().fold(
+                  () => SimplePostActionToolbar(post: rawPost),
+                  (post) => SzurubooruPostActionToolbar(post: post),
+                ),
+      ),
+      topRightButtonsBuilder: (currentPage, expanded, post) =>
+          GeneralMoreActionButton(post: post),
+      fileDetailsBuilder: (context, rawPost) => DefaultFileDetailsSection(
+        post: rawPost,
+        uploaderName: castOrNull<SzurubooruPost>(rawPost)?.uploaderName,
+      ),
+    );
+  }
+}
